backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
alembic # Để quản lý migrations (tùy chọn nâng cao, ban đầu có thể bỏ qua)